test: Run systemd in the test container
We want to have a test container that is as close as possible to a VM so we
don't have to spin VMs to test OS-level things the installer does, such a
installing packages or starting systemd services.

A step in that direction is to run systemd inside the container so we can
start/stop/enable systemd services.

In the test container systemd runs as pid 1. This needs a few things to work:
  - mounting the cgroup hierarchy of the running kernel inside the container
  - SYS_ADMIN capability
  - enabling the systemd docker support with env container=docker

We also remove a few systemd targets from the filesystem we don't really want
to execute in a container context.

We then execute tests with docker exec into that container. As we bind mount
the current git checkout inside the container so we can run tests in the
container.

We also re-use the trick to change the user to the uid/gid of the bind mount so
writes in that mount are owned by the correct user outside of the container.