charts: The cluster service needs permissions to entitlements
This is potentially temporary.

The entitlements we've made all refer to the weave-system namespace,
and it's part of the entitlement repo to look for an entitlement in
the weave-system namespace.

However, weave gitops core now runs in the flux-system namespace, so
it seemed easier to put the enterprise services in flux-system, and
just add this permission to read the existing entitlements.