charts: Turn on auth always
We must have auth turned on, or core doesn't work. It doesn't have to
be oidc, but it has to be enabled.

This may or may not be permanent (weaveworks/weave-gitops#1831) but as
things stand, all cluster queries in core uses impersonated
permissions and will fail unless auth is enabled.