Run snyk CLI checks on PRs
We've previously had the webhooks configured, which runs license &
dependency checks.

This replaces them with explicit CLI commands.

Why?

First, snyk isn't very good at go in general, and absolutely useless
when invoked through the webhooks - the webhooks just goes through the
go.sum file and reports everything in there as insecure, whether you
use it or not. The CLI uses go.mod instead, which actually works.

Second, this uses snyk-delta to alert on _new_ problems, not just "all
problems". This isn't perfect - it is subject to race conditions, when
a new upstream problem is discovered, but the monitor hasn't run since
- but it means a javascript engineer isn't stuck fixing go dependencies.