feat: Add MultiClusterFetcher (#654)
This commit adds a component which fulfills the Fetcher interface
defined in Core[1].

Likely this is an initial implementation. Since the code is very similar
to what the cluster-controller is doing, I did consider building an
in-memory cache there and then exposing it via an internal service, but
held off until we know what we want to do around efficiency here.
Querying the secrets and building the list with every call of `Fetch` may
not be what we want long-term. As it stands the caller of `Fetch` in
core will be caching the results (with a TTL), so the situation is not
too dire.

I did consider wiring this Fetcher to use a CRDLibrary[2] to List
clusters, but held off initially as, a) it did not save me anything
(indeed I do not need a map) and b) as illustrated above we may not want to
keep all of this mechanism here.

I have not wired in this Fetcher in the server yet, as there are some
final things which need to be done in core. Once those are merged and
the version of Core is bumped here I will PR to add in the wiring.

[1] https://github.com/weaveworks/weave-gitops/blob/main/core/clustersmngr/clustersmngr.go#L52-L56
[2] https://github.com/weaveworks/weave-gitops-enterprise/blob/main/cmd/clusters-service/pkg/clusters/clusters.go#L29-L33