Make the ui-cra image build UI inside container
Building the UI inside of a container saves a user from having to
install the right versions of node, yarn, and anything else. It also
makes it easy to produce the same build locally as we get from CI.

This achieves that by introducing 2 build targets, one for building
and one for running - this means that the build target can be cached,
but won't be part of the final image.

The build phase, in turn, has 2 phases, one for fetching node_modules
and one for building the prod js.

I've achieved this by teasing a couple of Makefile rules out of the
root Makefile and moved them into a new one in ui-cra - I've put
a wrapper in place in the root Makefile which I think should behave
the same so nobody would know.