Add first iteration of a CLI release process
This adds goreleaser, which is the tool core uses to make
releases. However, that's not the tool enterprise uses to make
releases, so the goreleaser configuration has explicitly disabled the
release bit.

The reason goreleaser is pretty useful is that it builds multi-arch
binaries, creates any archives we want, calculates checksums, uploads
them wherever we want, and automates generating packages for package
managers (currently only homebrew is enabled).

I've done some local testing to see that the goreleaser seems to
_mostly_ work, but there's more work to be done to make sure
 * releases are published To A Good Location™
 * all merges are published Somewhere Useful For Testing