Set enabled AuthMethods
The API for the auth server changed in
https://github.com/weaveworks/weave-gitops/pull/2522 from enabling
methods implicitly, based on configuration (e.g. whether the `oidc-auth`
secret existed) to explicitly. This was done to reduce the change of
accidental configuration or exploits that made use of nominally un-used
code-paths.

As part of this the signature changed for `NewAuthServerConfig` (the
addition of a list of enabled authentication methods). We add that here.

The PR also adds an encapsulated `InitAuthServer` function that, for
example, handles reading the `oidc-auth` secret. That function wasn't
used here as a different configuration flow is used here.