Upgrade BE to weave-gitops 0.9.2-rc.2 (#1211)
* Set enabled AuthMethods

The API for the auth server changed in
https://github.com/weaveworks/weave-gitops/pull/2522 from enabling
methods implicitly, based on configuration (e.g. whether the `oidc-auth`
secret existed) to explicitly. This was done to reduce the change of
accidental configuration or exploits that made use of nominally un-used
code-paths.

As part of this the signature changed for `NewAuthServerConfig` (the
addition of a list of enabled authentication methods). We add that here.

The PR also adds an encapsulated `InitAuthServer` function that, for
example, handles reading the `oidc-auth` secret. That function wasn't
used here as a different configuration flow is used here.

* Upgrade weave-gitops v0.9.2

* Pass runtimeNamespace to NewAuthServerConfig

A namespace needs to be passed to auth server config so it can look up
secrets in the correct namespace (by default, flux-system)

* Run linting

* Up to weave-gitops v0.9.2-rc.2

* Bump to flux 0.32 in acc. tests for OCI container support

* Pass runtimeNamespace through to server config

* Ooops, completely wrong node module url

* Pass runtime-ns through to ui-int server

Co-authored-by: Simon Howe <simon@weave.works>