Migrate gitops OSS version past weaveworks/weave-gitops#2900
The commit mentioned above removed `cmd/gitops/logger` and changed
`pkg/logger` to be a super-thin wrapper around logr. The behaviour is
mostly the same when you use `NewCLILogger`, but we no longer
need to initialize logr separately (we can just unwrap the one we
have), and `FakeLogger` is no more - to get a dummy logger, just call
`logr.Discard` and wrap it with `logger.From`.