Grab LDFLAGS from core so that gitops run works (#1732)
There are several things that core sets dynamically at build time -
last flux version, dashboard helm chart, dev-bucket image. New flags
were not added to the enterprise version, and actually keeping up with
the churn is boring.

This fetches the core Makefile from the commit that we're using, as
well as the helm chart from there as well (because for reasons that
aren't particularly great the core Makefile reads the version from the
chart). It then runs core's `echo-ldflags`, and uses those to link the
binary. It overrides the version, but otherwise leaves them alone.

In case that's not obvious from the change itself, I'm not
particularly proud of this - but I can't think of any other way to
grab the correct core 0.10 flags that don't require time travel, as we
didn't publish these anywhere else at the time.