default to Pipeline Namespace when querying target cluster
I had to change the MakeClustersManager func a little so that it's now
possible to pass a list of known clusters to it. If that list is
non-empty then the given cluster name is checked to be in that list
and if it isn't, an error is returned. This is closer to the behaviour
of the actual client and let me implement the unit test properly.

fixes #1748