Use new cluster interface
This is the enterprise side of the new Cluster interface, that
combines the old cluster with the old client factory.

It tries to avoid passing configs around - by making configs an
internal matter, things like timeouts, whether to wrap them in a
delegate client, etc, can all be configured from cluster to
cluster.Right now, all clusters are still created by the fetcher, so
there's still a single point of ownership, it's just outside the
clusters manager object.

There's an exception for the helm controller, where the
client generated from the controller-runtime has different settings
than the one we create, so in order to avoid (more) regressions I left
it.