Use new cluster interface
This is the enterprise side of the new Cluster interface, that
combines the old cluster with the old client factory.

It intentionally avoids passing configs around - it makes configs an
internal matter, so things like timeouts, whether to wrap them in a
delegate client, etc, can all be configured from cluster to
cluster. Right now, all clusters are still created by the fetcher, so
there's still a single point of ownership, it's just outside the
clusters manager object.