ui docker: Cache node_modules
I seem to spend a lot of time waiting for node to build
things. Granted, most of it I spend waiting for CRA to optimize
production builds (which this leaves unchanged), but I frequently also
have to wait for it to download yet another copy of node modules
because the gitops OSS revision changed.

This makes sure node modules is cached between runs even if
package.json changed, so that it can be reused the same way as my
local node_modules. This won't fix everything, but hopefully make
frontend development a little bit quicker.