fix: return with a list of error if appref kind if invalid
Kubernetes API and resources are case sensitive, because of that,
`helmrelease` as kind is invalid and it should be `HelmRelease`.

Based on discussions on Slack, we try to return as many information as
we can, so we don't return with an error if one of the apprefs is
incorrect, instead we return with a list of errors and don't add an
empty appref to the workloads list.

Related to #1924

References:
- https://github.com/weaveworks/weave-gitops-enterprise/issues/1924