Allow override of the downloadable kubeconfig.
This changes the way the kubeconfig is looked up when downloading.

If a secret is discovered in the same namespace as the Cluster with the
name "cluster-kubeconfig-override" then it will be provided instead.

This is assumed to be a KubeConfig, and will be validated as such.

Recommendation is to create one using the exec functionality and then
put it into the name secret, it will then be provided to every user if
they download the secret via the UI or via the API.

It also fixes up the user context with the user from the replaced authinfos.