fix: the new git provider didn't commit all required files
The root cause was a misused pointer reference in
`pkg/git/gitprovider.go`. For some reasons `go-git-providers` (GGP) expects a
reference to a string as filename. That wouldn't be an issue in most
cases, but we tried to convert WGE internal `CommitFile` to GGP's
`CommitFile`. In that loop we assigned a reference, but that reference
gets updated while we are in the loop:

```go
for _, f := range c.Files {
  adapted = append(adapted, gitprovider.CommitFile{
    Path:    &f.Path,
    Content: f.Content,
  })
}
```

As a result we end up with the same filename for all files, therefore we
commit only one file.

References:
* Previous PR: https://github.com/weaveworks/weave-gitops-enterprise/pull/2450
* Issue: https://github.com/weaveworks/weave-gitops-enterprise/issues/2413

Signed-off-by: Balazs Nadasdi <balazs@weave.works>