make reboot.sh more resilient to failures
I ran into an issue where the version of `clusterctl` in
`./tools/bin/` was too old which led to the `clusterctl init` step to
fail. However the script just glossed over that instead of bailing out
so I didn't notice and just went on trying to install WGE which
obviously failed.

Now, `reboot.sh` properly bails out on any error.

Signed-off-by: Max Jonas Werner <mail@makk.es>