Add facets support to query service
"Facets" refers to the possible values of a field -- e.g., for "kind",
the possible values are (say) "HelmRelease", "HelmChart",
"Kustomization", etc.

Adding these to the API lets the UI show all the options for filtering
on a field. Being able to do this is necessary for replicating the UI
of the existing Applications view, using the query API.

Using the keyword analyser makes the indexer look for exact matches,
rather than tokenising search terms. Here, it's made the default, but
it will be necessary later to set the analyser per field (so free text
search works on other fields).