Avoid logging problems with ObjectKind
In watching.go the .kinds field is included in a log; but it's a slice
of ObjectKind, and that type won't serialise to JSON because it has
func valued fields. This won't blow up the program, but it does result
in the logger printing a `kindError` entry with an error message
rather than the desired `kinds` entry.

Anyway: ignore the fields when serialising to JSON, by adding struct
tags.

Signed-off-by: Michael Bridgen <michael.bridgen@weave.works>