Filtering of the charts being added to the cache.
This adds functionality to the Helm watcher to parse an annotation on
the HelmRepository looking for a version constraint.

All charts added to this are checked against the constraint.

If the annotation doesn't exist, the existing behaviour is retained.

If the annotation is invalid, the filter is ignored and an error is
logged.