Cache downloaded (versioned) dependencies
I found myself running `make dependencies` a lot, to try and get all
the protoc generators accounted for. It downloads everything from
scratch, which takes ages. So: this commit changes it so it will
download a tool at a version once, and just copy it over to tools/bin/
if it already has it in the cache.

Signed-off-by: Michael Bridgen <michael.bridgen@weave.works>