Add protobuf generator plugins as dependencies
When I first tried

    make proto

it failed because it couldn't find a plugin it needed. I gather that
people have simply done `go install ...` to get these when (if) they
run into this problem.

This commit adds all the plugins it uses to dependencies.toml, so they
can be pinned to a version and repeatably give the same output.