Add support for access by resource name to query service
Kubernetes RBAC supports specifying access to individual resources
with `.resourceName`, and that field of each PolicyRule needs to be
stored.

ResourceName is usually empty, and we have to take care when
unstringifying it from the store, since

    join([], ",") -> ""

but

    split("", ",") -> [""]

which would represent a requirement for an empty name. Therefore: an
empty string now unstringifies to a zero-value.