Use the Kubernetes RBAC machinery
This replaces the local implementation for evaluating RBAC rules with
a shim for the Kubernetes implementation of the same. This involves:

 - ... retrieving Role and RoleBinding values from the store and using
   them directly, rather than flattening them into the intermediate
   AccessRule. The Kubernetes machinery calculates the effective rules
   given (Cluster) Role and RoleBinding, and it is impossible to
   recover those, or otherwise to do the calculation of effective
   rules, with AccessRules.

 - ... adapting between types in `models` and the Kubernetes RBAC
   types. In the case of API objects, this involves copying field
   across, since they are structs. For other types, there are wrappers
   or casts.

In Kubernetes, the mode of use is that the user and resource are
given, and you must look up the relevant rules and evaluate
them. Here, the access pattern is that the user is given, then you
need to evaluate access for objects which can be in several clusters
and namespaces.

I have kept it simple (and similar to the prior implementation) by
retrieving Role and RoleBinding objects for all clusters and
namespaces, then running through those when asked by the Kubernetes
machinery while it figures out which rules apply. The relevant rules
are calculated at most once per cluster. There's a lot of room for
improvement, since that still runs through all the objects, each time.

I have kept the debug API DebugAccessRules and supporting bits
(notably `accesschecker#RelevantRulesForUser`), for now, since these
have UI which depend on them.

Last caveat: it's still necessary to translate between the kind given
for an object, and the resource expected by the RBAC
machinery. Previously, a map of resource to kind was constructed, and
(effectively) the resource part of the rule was looked up in it to get
a kind. I have switched that around: the kind of an object is looked
up, to compare to the resource field of each rule -- simply because
object.GetResource() is implemented here, and that's my opportunity to
interpose the lookup.

Signed-off-by: Michael Bridgen <michael.bridgen@weave.works>