Benchmark RBAC access control
This is an attempt at benchmarking the RBAC code. It tries to build a
representative model -- a modest number of clusters, roles and
rolebindings (most of which do not pertain to the user or objects in
question), and a list of objects to check. Then it uses
rbac.Authorizer in the way it's used for queries (memoising the
instantiated authorisation predicate per cluster).

This should give us a baseline with which to judge any
improvements. It will not necessarily be a good guide to relative
effectiveness of improvements, since results are sensitive to how the
model is constructed.

Signed-off-by: Michael Bridgen <michael.bridgen@weave.works>