Caching for rolebindings and roles
From the store, we are given all Role objects and all RoleBinding
objects. To adapt to the Kubernetes RBAC package, we need to provide

 - GetRole
 - ListRoleBindings

and the ClusterRole and ClusterRoleBinding analogues. Since we're
given a name in GetRole, the implementation can stop when it sees that
namespace and name; but ListRoleBindings must scan all the
RoleBindings to find all those in the namespace.

With that in mind, this commit uses different ways of avoiding looking
at the whole list of roles or rolebindings more than once:

For GetRole, an index of name to position in the list is kept. The
index is checked before starting, in case the name has been seen
already. Otherwise, the list is scanned, recording the position of
each Role, until it sees the named object. It keeps track of where in
the list it got up to, so the next invocation can start there.

For ListRoleBindings, the list is sorted into namespaces before any
methods are called. Trying to do this incrementally is complicated;
and besides, as noted above, any invocation will necessarily scan the
whole list.

These both duplicate work done for other clusters, since each
implementation gets all the objects rather than just those for its
cluster.

Signed-off-by: Michael Bridgen <michael.bridgen@weave.works>