Avoid mapping subresources to kinds
Kubernetes RBAC deals with resources rather than kinds, but what we
want to check are objects with kinds. The solution to date is to map
kinds to resources (previously, resources to kinds), and to obtain
that map using a Discovery client.

It turns out that if you ask the discovery client
`ServerGroupsAndResources()`, for any kind there is usually at least
two resources in the result -- the "main" one e.g., "helmreleases",
and a subresource for the status, e.g., "helmreleases/status". This
complicates things because the latter will overwrite the first.

Previously (resource->kind), you'd get

    {
      "helmreleases": "HelmRelease",
      "helmreleases/status": "HelmRelease"
    }

but mapping the other way, you get

    {
      "HelmRelease": "helmreleases/status"
    }

and none of the PolicyRules, which use "helmreleases" as the resource,
match.

Happily there's an easy fix: if you use `ServerPreferredResources()`
you get just the main resources.

Signed-off-by: Michael Bridgen <michael.bridgen@weave.works>