Disentangle Collector/Watcher implementations
At present the Collector implementation is initialised with a bunch of
values ((e.g., the object kinds to watch) and some callback funcs
(`NewWatcherFunc` and `ProcessRecordFunc`). The callback funcs are
called with the values as supplied (object kinds) and some values
particular to that call (e.g., the cluster that the `Watcher` is for).

This design ties together the implementation of the collector, the
watchers, and the callback functions, since they all have to agree on
what values will be used by the callback functions.

This commit removes the "fixed" arguments for `NewWatcherFunc` from
its signature; implementations can get what they need from their
lexical closure instead. The collector implementation doesn't need to
know about them, so they can be removed from `CollectorOpts`. Now, the
way a watcher works can be changed without the collector needing to
be changed in sympathy.

Similarly, we want to be free to change what the watcher does with the
objects it finds, without needing to change the collector; so, rather
than the collector constructing a channel and running a callback, let
the caller do that (see `NewRoleCollector` and `NewObjectsCollector`).

Lastly, the RoleCollector and ObjectsCollector structs don't do
anything other than hold a Collector (and in fact, seem to have fields
that aren't initialised or used at all); so, they get deleted.

`NewRoleCollector` and `NewObjectsCollector` are retained, since they
encode the specifics of which kinds of object to watch for, and what
to do with results. (I may end up reinstating the structs, so as to
have some shutdown protocol since there are goroutines. But for now,
simplify.)

NB removed lines in tests represent objects that no longer need to be
constructed as part of setting things up -- no _assertions_ are
removed.

Signed-off-by: Michael Bridgen <michael.bridgen@weave.works>