Test degenerate case of collector shutdown
That is: check that if you start a collector with nothing to do, then
shut it down, no goroutines are left running.

This is not the most important test case of this kind, since
collectors have the same lifetime as the process; but, it's a place to
start with these kinds of checks.

Signed-off-by: Michael Bridgen <michael.bridgen@weave.works>