Exit collector subscription goroutine on shutdown
A collector starts a goroutine to monitor the Updates channel from a
clusters manager. To shutdown gracefully, that goroutine needs to
exit. There's a few ways to do this:

 - closing the Updates channel would let the goroutine (which just
   reads from it) exit; but, the Updates channel doesn't belong to us!
   Calling `clusterswatcher.Unsubscribe()` happens to close the
   channel, but I don't think we can rely on that (although it should
   be done, otherwise upstream might be blocked...).

 - using a cancellable context, then selecting between ctx.Done() and
   Updates. We don't need a context for anything inside the loop,
   though, so...

 - construct a channel just for that purpose.

The sync.WaitGroup is there so .Stop() can shutdown synchronously. I'm
not sure it's necessary for it to be synchronous, at present, but I
don't think it hurts either, and it's easier to understand for the
caller.

Signed-off-by: Michael Bridgen <michael.bridgen@weave.works>