Unsubscribe to clusters when stopping collector
When you use ClustersManager.Subscribe(), you get a handle
*ClustersWatcher, with a channel holding updates. If you don't use the
handle to Unsubscribe() when you stop dequeueing updates, the manager
will try to enqueue, and block. So: call Unsubscribe.

Unfortunately, it's impossible to use `fakeclustermngr` for the tests
here, because the handle you get back from `Subscribe()` is a concrete
type with an unexported field, and any value constructed by a fake
will NPE when you call `Unsubscribe()` on it.

So: this commit introduces interfaces (in collector/clusters) to
represent the requirements of the collector objects, meaning they can
have fake objects. `Subscriber.Subscribe()` returns another interface,
also with a fake. A wrapper `MakeSubscriber` is given to adapt
`clustersmngr.ClusterManager` to the local interface.

Signed-off-by: Michael Bridgen <michael.bridgen@weave.works>