Combine collector stop tests
There's already a test for stopping the default collector, so just
test for goroutine leaks there. I removed the loop over test cases
because the collector is started outside the loop and stopped _inside_
it, which wouldn't work if there were more than one test; and, rather
than moving the stop inside the loop it was simpler to not loop.

Signed-off-by: Michael Bridgen <michael.bridgen@weave.works>