Impersonate as soon as possible
This commit moves the construction of the impersonating client config
to the Collector implementation, with the goal of having fewer points
where it could be forgotten.

In principle the SA could just be passed through to NewWatcher by the
caller of NewCollector; however, that would leave open the
possibilities of a caller of NewCollector forgetting to do this, or
the implementation of the watcher not bothering to use it. This way,
it's part of the protocol for the Collector, and no other component
gets a chance to muck it up.

Signed-off-by: Michael Bridgen <michael.bridgen@weave.works>