Use a workqueue to process cluster watches
At present, when the collector fails to watch a cluster, that's it --
the cluster will never be watched. So as to be able to handle retries,
this commit interposes a client-go workqueue (same as that used in
controller-runtime) between the notification of a new cluster and it
being watched. This way, if the watch fails, it can be retried.

Since the watchers are now run in goroutines managed by the collector,
it can see when one fails and retry it.

There is a bit of delicate (brittle?) tip-toeing around clusters that
are invalid; e.g., are missing a client config. These return an error
before starting a watcher. I have chosen to mark them as failed, and
not requeue them. If they come in on the clustersmngr channel again,
they will be retried then.

This commit adds a watcher status "starting", for the period when a
cluster can be started, but hasn't been yet. This makes sure that if a
cluster is recorded at all, it has a non-empty status.

Signed-off-by: Michael Bridgen <michael.bridgen@weave.works>