Use callback for processing object updates
Make the reconciler take a callback, rather than a channel. Why? A
callback in more general -- if you want backpressure or
synchronisation, you can always make the func push to a chan. This
also gives the opportunity to retry transactions, which was not
implemented before and anyway would be difficult using channels.

The "objects channel" is still passed down as far as the manager, in
this commit. It's fairly easy to remove it from the Options structs
and constructor funcs further up, but it would be better to simplify
these out entirely.

Signed-off-by: Michael Bridgen <michael.bridgen@weave.works>