Make watcher interface look like manager.Manager
The watcher implementation is mostly a wrapper around the controller
manager, so exists to

 - be something you can start and stop; and,
 - hold some values before passing them on to a NewManager callback

There's no particular need to do this dance -- it might be useful to
use a fake controller-runtime manager in tests, but all those tests
really cover is whether the scaffolding supports itself.

This commit makes space for removing Watcher altogether, by changing
the signature returned by NewWatcherFunc to match
manager.Manager#Start(ctx).

The consequences are:

 - the watchingCollector starts a goroutine for each watcher

This is because manager.Manager#Start blocks until its context is
canceled, or it fails. There was a goroutine already -- this just
moves it up the collector, and obviates some synchronisation that was
necessary to adapt between `Start()` and `Stop()` and `Start(ctx)`.

Since watcher failures are now visible to the collector, it can take
action on them, like restarting the watcher.

 - the watchingCollector keeps track of each watcher's status

This expands the statuses to "started"|"stopped"|"failed".

Signed-off-by: Michael Bridgen <michael.bridgen@weave.works>