Reinstate hook for deleting cluster contents
Previously, when the watcher implementation was asked to stop, it sent
a "delete all" message to the object transaction channel. The effect
of this was to delete all items from the cluster in question, from the
store and index.

Now that manager.Manager is constructed directly, this finalisation
has to be done somewhere else. A reasonable place is the collector,
since that deals with the watcher lifecycle; e.g., it's also the
called of `NewWatcherFunc`.

Signed-off-by: Michael Bridgen <michael.bridgen@weave.works>