Use Starter interface for collector
It's handy to be able to cancel a context to stop goroutines, rather
than having to remember to call stop on each one; and, there's less
synchronisation needed. So: use the `Start(context.Context) error`
method for Collector as well, rather than Start/Stop.

In this case, the loop in watchingCollector doesn't need to be its own
goroutine, which makes it a bit easier to think about.

The trade-off is that you need to explicitly construct a goroutine for
the (blocking) Start, if you do want it to run in another goroutine.

Signed-off-by: Michael Bridgen <michael.bridgen@weave.works>