Adjust tests for asynchronous Start(ctx)
Previously, some tests in watching_test.go relied on the
watchingCollector having synchronously processed existing clusters
before it returns from Start(). This is a bit of a coincidence; it's
also no longer the case, because the collector is started in its own
goroutine. So: use `g.Eventually` to check processing has happened.

Signed-off-by: Michael Bridgen <michael.bridgen@weave.works>