Make collector stop test non-flakey
I wanted to test when you stop the collector, all the goroutines it
starts will exit. But, counting the number of goroutines isn't
reliable for this, since there may be running goroutines from other
tests hanging around. (Quite often it was failing because the number
after stopping was _less_ than the number before starting ...).

To avoid this particular flake, this commit reduces the strength of
the test to just checking that the collector returns from
`Start(ctx)`.

Signed-off-by: Michael Bridgen <michael.bridgen@weave.works>