Add `unstructured` field to query service objects.
In an effor to allow for any kubernetes object to be captured by the objectscollector and shown, we also need to allow for the entire raw document to be saved in the database. This allows for customizable UIs on the front end to be built with a for a specific "kind".