Ignore docs/ and similar paths for test workflow
The tests workflow just runs unit tests and integration tests (make
targets `integration-tests` and `unit-tests-with-coverage`). I've
added some fairly conservative paths to the GitHub workflow under
on.paths-ignore, so that PRs with _just_ changes to docs aren't
contingent on the tests. This saves some waiting, and some resources.

 - docs/** # these are not involved in the tests
 - bin/** # scripts that are not involved in the tests

In general we _do_ want to run tests if it's the workflow itself
that's changed, so I've bene more specific in `.github/` and just
ignored the templates.

Signed-off-by: Michael Bridgen <michael.bridgen@weave.works>