Add granular control of query service components
We want to be able to enable and disable the Explorer and query service backend for different parts of the UI. Before this, it was a global on-off switch handled via feature flags. This commit adds a dedicated enpoint for us to hit to enable or disable components via the helm values file.